# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Striker EVM API
  version: 1.0.0
  summary: Contrato v1 del API de Striker EVM (Earned Value Management).
  description: |
    API REST de **Striker EVM**, herramienta interna para que los líderes de proyecto registren el
    avance de sus actividades y obtengan, en tiempo real, los indicadores de Valor Ganado (EVM)
    por actividad y consolidados por proyecto.

    Este documento es el **punto de sincronización entre backend y frontend** (módulo M1 de
    `docs/ARQUITECTURA.md`). El backend debe exponer exactamente estas rutas, esquemas y códigos;
    el frontend genera su cliente tipado a partir de este archivo y usa las fixtures de
    `docs/api/fixtures/` como mocks. Cualquier cambio al contrato se hace en un Pull Request propio.

    ## Convenciones

    - Prefijo de todas las rutas: `/api/v1`.
    - JSON en `camelCase`.
    - **Dinero** (BAC, AC, PV, EV, CV, SV, EAC, VAC): número con exactamente **2 decimales**,
      redondeo *half up*.
    - **Índices** (CPI, SPI): número con exactamente **4 decimales**, redondeo *half up*.
    - **Porcentajes**: escala 0–100 tal como los ingresa el usuario, hasta 2 decimales.
    - Indicador **no calculable** (división por cero): valor `null`, estado `NOT_APPLICABLE` y una
      explicación en `notes`. Nunca se inventan valores (ni 0, ni 1, ni infinito).
    - Los indicadores **no se persisten**: se calculan en cada consulta a partir de los datos crudos.
...

    ## Autenticación y roles

...

...
    |---|---|
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/activities:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Actividades]
      operationId: listActivities
      summary: Listar actividades de un proyecto
      description: |
        Devuelve las actividades del proyecto con sus **datos crudos** (presupuesto, porcentajes y
        costo real), sin indicadores. Para los indicadores usar `GET /projects/{projectId}/evm`.
        Disponible para ambos roles. Un proyecto sin actividades responde `200` con `[]`.
      responses:
        '200':
          description: Lista de actividades ordenada por fecha de creación ascendente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Actividades]
      operationId: createActivity
      summary: Crear actividad
      description: |
        Crea una actividad dentro del proyecto. Disponible para ambos roles con esta regla de
        responsable:

        - `REGISTRAR`: `ownerId` es opcional y, si se envía, debe ser su propio id; la actividad
          queda a su nombre.
        - `REVIEWER`: `ownerId` es **obligatorio** y debe corresponder a un usuario existente.

        Un `ownerId` ausente (REVIEWER) o inexistente responde `400` con código
        `VALIDATION_ERROR`; un `ownerId` ajeno (REGISTRAR) responde `403` con código `FORBIDDEN`,
        porque la política de responsables prohíbe la acción en lugar de corregirla en silencio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityInput'
      responses:
        '201':
          description: Actividad creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /projects/{projectId}/activities/{activityId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/ActivityId'
    put:
      tags: [Actividades]
      operationId: updateActivity
      summary: Editar actividad
      description: |
        Reemplaza los datos de la actividad. `REVIEWER` puede editar cualquier actividad y
        reasignar `ownerId`; `REGISTRAR` solo puede editar las actividades de las que es
        responsable y no puede cambiar el responsable (cualquier otra combinación responde `403`).
        La actividad debe pertenecer al proyecto indicado; si no, responde `404`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityInput'
      responses:
        '200':
          description: Actividad actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Actividades]
      operationId: deleteActivity
      summary: Eliminar actividad
      description: Elimina la actividad. `REVIEWER` puede eliminar cualquiera; `REGISTRAR` solo las propias (`403` en caso contrario).
      responses:
        '204':
          description: Actividad eliminada. Sin cuerpo.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT obtenido en `POST /auth/login`. Enviar como `Authorization: Bearer <token>`.
        Claims: `sub` (id del usuario), `role` (`REGISTRAR` | `REVIEWER`), `exp` (8 h).
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Identificador (UUID) del proyecto.
      schema:
        type: string
        format: uuid
      example: 22222222-2222-4222-8222-000000000001
    ActivityId:
      name: activityId
      in: path
      required: true
      description: Identificador (UUID) de la actividad. Debe pertenecer al proyecto de la ruta.
      schema:
        type: string
        format: uuid
      example: 33333333-3333-4333-8333-000000000002
  responses:
    BadRequest:
      description: Cuerpo inválido o regla de negocio de entrada incumplida (`VALIDATION_ERROR`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: La solicitud contiene campos inválidos
            details:
              - field: plannedProgressPercent
                message: plannedProgressPercent must be between 0 and 100
    Unauthorized:
      description: Token ausente, inválido o expirado (`UNAUTHORIZED`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Se requiere un token de acceso válido
            details: []
    Forbidden:
      description: El usuario está autenticado pero su rol o su relación con el recurso no le permiten la acción (`FORBIDDEN`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: Solo un usuario con rol REVIEWER puede crear proyectos
            details: []
    NotFound:
      description: El recurso indicado en la ruta no existe (`NOT_FOUND`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Proyecto 22222222-2222-4222-8222-000000000009 no encontrado
            details: []
    InternalError:
      description: Error inesperado del servidor (`INTERNAL_ERROR`). El cuerpo nunca expone detalles internos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: INTERNAL_ERROR
            message: Error interno del servidor
            details: []
  schemas:
...
    Percent:
      type: number
      format: double
      minimum: 0
      maximum: 100
      description: Porcentaje de avance en escala 0–100, hasta 2 decimales.
      examples: [50.00, 100.00, 0]
    UserSummary:
      type: object
      description: Referencia resumida a un usuario (responsable de actividad o creador de proyecto).
      required: [id, fullName]
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
      examples:
        - id: 11111111-1111-4111-8111-000000000002
          fullName: Carlos Registrador

    # ------------------------------------------------------------------ errores
    ErrorCode:
      type: string
      description: |
        Catálogo de códigos de error del API. `HTTP_ERROR` es el código de los estados que
        produce el framework y ningún caso de uso levanta (por ejemplo `405 Method Not Allowed`).
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - HTTP_ERROR
        - INTERNAL_ERROR
    ErrorDetail:
      type: object
      description: Detalle de un problema puntual, normalmente asociado a un campo del cuerpo.
      required: [message]
      properties:
        field:
          type: [string, 'null']
          description: Nombre del campo (`camelCase`) al que aplica el detalle, o `null` si es general.
        message:
          type: string
          description: Descripción legible del problema.
        type:
          type: [string, 'null']
          description: |
            Identificador técnico del error de validación (por ejemplo `int_parsing`). Solo
            aparece cuando la falla es de forma del cuerpo, no de una regla de negocio.
      examples:
        - field: budgetAtCompletion
          message: budgetAtCompletion must be greater than 0
    Error:
      type: object
      description: Cuerpo uniforme de todas las respuestas 4xx/5xx.
      required: [code, message, details]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Mensaje legible para el usuario final.
        details:
          type: array
          description: Detalles adicionales (uno por campo inválido en `VALIDATION_ERROR`); vacío en los demás casos.
          items:
            $ref: '#/components/schemas/ErrorDetail'
      examples:
        - code: VALIDATION_ERROR
          message: La solicitud contiene campos inválidos
          details:
            - field: plannedProgressPercent
              message: plannedProgressPercent must be between 0 and 100

    # ------------------------------------------------------------------ sistema
...
    ActivityMeasures:
      type: object
      description: |
        Datos crudos de avance y costo de una actividad; son la entrada del cálculo EVM.
        `budgetAtCompletion` debe ser mayor que 0 (un presupuesto cero se rechaza con `400`);
        `actualCost` puede ser 0 y los porcentajes pueden ser 0 (estados legítimos).
      required: [budgetAtCompletion, plannedProgressPercent, actualProgressPercent, actualCost]
      properties:
        budgetAtCompletion:
          type: number
          format: double
          exclusiveMinimum: 0
          maximum: 999999999999.99
          description: BAC. Presupuesto total planificado de la actividad al 100 %. Debe ser > 0. Dos decimales.
        plannedProgressPercent:
          $ref: '#/components/schemas/Percent'
          description: Porcentaje de avance planificado a la fecha de corte (0–100).
        actualProgressPercent:
          $ref: '#/components/schemas/Percent'
          description: Porcentaje de avance real a la fecha de corte (0–100).
        actualCost:
          type: number
          format: double
          minimum: 0
          maximum: 999999999999.99
          description: AC. Costo real incurrido hasta la fecha. Debe ser >= 0. Dos decimales.
      examples:
        - budgetAtCompletion: 40000.00
          plannedProgressPercent: 50.00
          actualProgressPercent: 40.00
          actualCost: 20000.00
    ActivityInput:
      description: |
        Cuerpo de creación y edición de una actividad (POST y PUT).

        `ownerId`: para `REGISTRAR` es opcional (se asigna a sí mismo; si lo envía debe ser su propio
        id); para `REVIEWER` es obligatorio y debe existir. Incumplirlo responde `400`.
      allOf:
        - $ref: '#/components/schemas/ActivityMeasures'
        - type: object
          required: [name]
          properties:
            name:
              type: string
              minLength: 1
              maxLength: 120
              description: Nombre de la actividad.
            ownerId:
              type: [string, 'null']
              format: uuid
              description: Id del usuario responsable. Ver reglas por rol.
      examples:
        - name: Desarrollo
          ownerId: 11111111-1111-4111-8111-000000000003
          budgetAtCompletion: 40000.00
          plannedProgressPercent: 50.00
          actualProgressPercent: 40.00
          actualCost: 20000.00
    Activity:
      description: Actividad persistida con sus datos crudos (sin indicadores).
      allOf:
        - $ref: '#/components/schemas/ActivityMeasures'
        - type: object
          required: [id, projectId, name, owner, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            projectId:
              type: string
              format: uuid
            name:
              type: string
            owner:
              $ref: '#/components/schemas/UserSummary'
              description: Usuario responsable de la actividad.
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
      examples:
        - id: 33333333-3333-4333-8333-000000000002
          projectId: 22222222-2222-4222-8222-000000000001
          name: Desarrollo
          owner:
            id: 11111111-1111-4111-8111-000000000003
            fullName: Ana Registradora
          budgetAtCompletion: 40000.00
          plannedProgressPercent: 50.00
          actualProgressPercent: 40.00
          actualCost: 20000.00
          createdAt: '2026-09-08T12:00:00Z'
          updatedAt: '2026-09-08T12:00:00Z'

    # ------------------------------------------------------------------ reporte EVM
...